Share Ecmascript HMR chunk versioning, diffing, and merging between browser and node
