--- turbopack/crates/turbopack-browser/src/ecmascript/merged/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod content;
-pub(crate) mod merger;
-pub(crate) mod update;
-pub(crate) mod version;
