--- turbopack/crates/turbopack-nodejs/src/ecmascript/node/merged/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod content;
-pub(crate) mod merger;
-pub(crate) mod update;
-pub(crate) mod version;
