+++ turbopack/crates/turbopack-nodejs/src/ecmascript/node/mod.rs
@@ -1,6 +1,3 @@
 pub(crate) mod chunk;
 pub(crate) mod content;
 pub mod entry;
-pub(crate) mod merged;
-pub(crate) mod update;
-pub(crate) mod version;
